+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_utils.py
@@ -3,7 +3,10 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+
 def get_http_request_kwargs(kwargs):
     http_request_keywords = ["params", "headers", "json", "data", "files"]
-    http_request_kwargs = {key: kwargs.pop(key, None) for key in http_request_keywords if key in kwargs}
+    http_request_kwargs = {
+        key: kwargs.pop(key, None) for key in http_request_keywords if key in kwargs
+    }
     return http_request_kwargs