+++ sdk/schemaregistry/azure-schemaregistry/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0 (Unreleased)
+## 1.0.0 (2021-11-09)
 
 **Note:** This is the first stable release of our efforts to create a user-friendly and Pythonic client library for Azure Schema Registry.
 
@@ -10,7 +10,8 @@
   - `register_schema`: Store schema in the service by providing schema group name, schema name, schema format and schema definition.
   - `get_schema`: Get schema definition and its properties by schema id.
   - `get_schema_properties`: Get schema properties by providing schema group name, schema name, schema format and schema definition.
-- `SchemaProperties` has the following instance variables: `id` and `format`.
+- `SchemaProperties` has the following instance variables: `id` and `format`:
+  - The type of `format` has been changed from `str` to `SchemaFormat`.
 - `Schema` has the following properties: `properties` and `definition`.
 - `SchemaFormat` provides the schema format to be stored by the service. Currently, the only supported format is `Avro`.
 - `api_version` has been added as a keyword arg to the sync and async `SchemaRegistryClient` constructors.