+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.6.0b1 (2021-11-5)
+## 12.6.0b1 (2021-11-08)
 **New features**
 - Added support for batch deleting files using the `delete_files()` method from a `FileSystemClient`
 - Account level SAS tokens now support two new permissions: